* dev:
  handle test cases impact
  Update Set Lib testcase
  add test cases for kernel core
  no need to test this
  Change Set LIB to background job
  recover back error delete code
  add cases todo list and remove completed todo comments
  Change interface for filling block.
  Init LastIrreversibleBlockJob
  todo
  support smart contract initialize method
  support initialize method call after deploying a smart contract
  ci
  change method params in bridge context
  record block info
  fix tx pool record
  management build passed

# Conflicts:
#	AElf.Kernel.SmartContract.Sdk.Abstracts/ISmartContractBridgeContext.cs
#	AElf.Kernel.SmartContract/HostSmartContractBridgeContext.cs
#	AElf.OS.Core/Node/Application/IOsBlockchainNodeContextService.cs

+++ AElf.Management.Website/Controllers/TransactionController.cs
@@ -20,10 +20,10 @@ namespace AElf.Management.Website.Controllers
 
         [HttpGet]
         [Route("poolsize/{chainId}")]
-        public async Task<ApiResult<ulong>> PoolSize(string chainId)
+        public async Task<ApiResult<int>> PoolSize(string chainId)
         {
             var result = await _transactionService.GetPoolSize(chainId);
-            return new ApiResult<ulong>(result);
+            return new ApiResult<int>(result);
         }
 
         [HttpGet]
